Added simple version/checksum to commit info.
